refactor: result관련 파일 result패키지로 이동
